requirements.txt
chainer >= 1.21
fire
matplotlib
pillow
